requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
distlib==0.3.1
Django==3.2
djangorestframework==3.12.4
filelock==3.0.12
importlib-metadata==3.10.1
pipenv==2020.11.15
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
virtualenv==20.4.3
virtualenv-clone==0.5.4
zipp==3.4.1
